requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
kombu==5.3.5
pillow==10.2.0
prompt-toolkit==3.0.43
pydantic==2.6.0
pydantic_core==2.16.1
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
qrcode==7.4.2
segno==1.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
